ARRAY_LENGTH macro returns int

+++ src/util.h
@@ -2,6 +2,6 @@
 #define UTIL_H
 
 // TODO: Cast to int
-#define ARRAY_LENGTH(arr) (sizeof(arr) / sizeof(*(arr)))
+#define ARRAY_LENGTH(arr) ((int)(sizeof(arr) / sizeof(*(arr))))
 
 #endif // UTIL_H
